requirements.txt
# AI学习项目依赖包
# 基础依赖
python-dotenv==1.0.0
requests==2.31.0

# LangChain生态
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0

# 向量数据库
faiss-cpu==1.7.4
chromadb==0.4.22

# 通义千问API
dashscope==1.17.0

# 可选API客户端
openai==1.6.1

# 数据处理
numpy==1.24.3
pandas==2.0.3

# Web框架
fastapi==0.104.1
uvicorn==0.24.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 开发工具
jupyter==1.0.0
streamlit==1.28.1